% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annotations_to_sf2}
\alias{annotations_to_sf2}
\title{Convert features to sf objects (updated)}
\usage{
annotations_to_sf2(
  p,
  feature = c("box", "cam", "ray", "ves", "pith"),
  col = NULL,
  r = 0.4424
)
}
\arguments{
\item{p}{Picture file output from \code{read_svg} function}

\item{feature}{Type of feature to be converted.}

\item{col}{Hexcode color used for the feature. Typically for
box=c("#44330BFF","#00FF00FF"),
cam="#00FFFFFF",
ray="#FF00FFFF",
ves="#FFFF00FF",
pith="#543D89FF"}

\item{r}{Size (in micrometers) of each pixel.}
}
\value{
A \code{sf} object with the annotated features.
}
\description{
Convert annotated features from the \code{Picture} object into \code{sf} objects
for spatial analyses.
}
\details{
This is an updated version of the function that works for later versions of \code{rsvg} and \code{grImport2} packages (new way of storing the svg data).
}
