% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colorID2}
\alias{colorID2}
\title{Check color codes on annotated SVG file (new SVG version)}
\usage{
colorID2(p)
}
\arguments{
\item{p}{Picture file output from \code{read_svg} function}
}
\value{
An object of class \code{Picture} from the \code{grImport2} package.
}
\description{
This function gives the colors of various annotated objects in an SVG file
as well as the count of objects in each color.
}
