% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{feature_numbers}
\alias{feature_numbers}
\title{Count feature types}
\usage{
feature_numbers(
  p,
  box_col = c("#44330BFF", "#00FF00FF"),
  cam_col = "#00FFFFFF",
  ray_col = "#FF00FFFF",
  ves_col = "#FFFF00FF",
  pith_col = "#543D89FF"
)
}
\arguments{
\item{p}{Picture file output from \code{read_svg} function}

\item{box_col}{hexcode color used for the bounding box}

\item{cam_col}{hexcode color used for the cambium line}

\item{ray_col}{hexcode color used for the ray lines}

\item{ves_col}{hexcode color used for the vessels}

\item{pith_col}{hexcode color used for the pith polygon}
}
\value{
A matrix of feature counts.
}
\description{
Gives the number of different feature types in the annotated SVG file based on color codes.
}
