% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grid_spacing}
\alias{grid_spacing}
\title{Compute grid spacing for tissue fraction analysis}
\usage{
grid_spacing(diameter = NULL, min_overlaps = 400)
}
\arguments{
\item{diameter}{The diameter of the section in microns, measure with a line tool and put length here}

\item{min_overlaps}{Define the minimum number of overlapping grid crosses (default is 400)}
}
\value{
The grid spacing (in microns) to set in QuPath.
}
\description{
Compute grid spacing for tissue fraction analysis.  Based on the area of the sample and a desired minimum number of grid points, this will provide the size of grid spacing.
}
