% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{group_size}
\alias{group_size}
\title{Group size}
\usage{
group_size(vsf, thresh = 33.26)
}
\arguments{
\item{vsf}{\code{sf} object of vessel polygons}

\item{thresh}{Distance threshold (in microns) to idenfity grouped vessels.}
}
\value{
A vector of group size for each vessel.
}
\description{
Compute the size of the group that each vessels belongs to.
}
