% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwar-package.R
\docType{package}
\name{qwar-package}
\alias{qwar}
\alias{qwar-package}
\title{Quantitative wood anatomy analyses in R}
\description{
Perform analysis of quantitive wood anatomy based on annotated SVG files.
}
\details{

}
\author{
\strong{Maintainer}: Robert Muscarella \email{bob.muscarella@gmail.com} (\href{https://orcid.org/0000-0003-3039-1076}{ORCID})

}
\keyword{anatomy}
\keyword{quantitative}
\keyword{wood}
