% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_svg}
\alias{read_svg}
\title{Load an annotated SVG file}
\usage{
read_svg(svgfile)
}
\arguments{
\item{svgfile}{Character string name of the input file (ending in .svg)}
}
\value{
An object of class \code{Picture} from the \code{grImport2} package.
}
\description{
This function loads an annotated SVG file, output from QuPath.
It can take a minute or so, depending on the file size.
}
