% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{thumbnail_check}
\alias{thumbnail_check}
\title{Create a thumbnail image if annotations}
\usage{
thumbnail_check(
  outfile = NULL,
  bsf = NULL,
  csf = NULL,
  rsf = NULL,
  vsf = NULL,
  psf = NULL,
  vstat = "area"
)
}
\arguments{
\item{outfile}{Character name of output file (end in .pdf), optionally including the path.}

\item{bsf}{\code{sf} object of bounding box}

\item{csf}{\code{sf} object of cambium line}

\item{rsf}{\code{sf} object of ray line(s)}

\item{vsf}{\code{sf} object of vessel polygons}

\item{psf}{\code{sf} object of pith polygon}

\item{vstat}{Attribute from vessel polygons to use in plotting.}
}
\value{
Saves a PDF file showing the output annotations.
}
\description{
Quickly create a thumbnail image to confirm annotations are being
properly processed by \code{qwar} functions
}
