% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ves_characteristics}
\alias{ves_characteristics}
\title{Vessel characteristics on a whole sample}
\usage{
ves_characteristics(vsf, diam = "Dcircle")
}
\arguments{
\item{vsf}{\code{sf} object of vessel polygons}
}
\value{
A data.frame with characteristics of vessels included in each distance bin.
Includes vessel grouping indices, mean and median vessel area, kurtosis, skewness,
hydraulically-weighted diameter (Dh), total theoretical conductivity (Kh_total),
and mean theoretical conductivity (Kh_mean).
}
\description{
Get characteristics of vessels in an entire (unbinned) sample.
}
