% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ves_diam_circle}
\alias{ves_diam_circle}
\title{Circle vessel diameter}
\usage{
ves_diam_circle(vsf, verbose = F)
}
\arguments{
\item{vsf}{\code{sf} object of vessel polygons}
}
\value{
A vector of cirular diameter (in microns) for each vessel.
}
\description{
Compute the diameter of vessels based on a circle with equal area.
}
