% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ves_diam_median_radius}
\alias{ves_diam_median_radius}
\title{Vessel diameter based on median radius}
\usage{
ves_diam_median_radius(
  vsf,
  parallel = TRUE,
  cores = detectCores(logical = FALSE)
)
}
\arguments{
\item{vsf}{\code{sf} object of vessel polygons}

\item{parallel}{Logical if processing should be done in parallel
(implemented but of questionably utility).}

\item{cores}{Number of cores to use for parallel processing.}
}
\value{
A vector of diameter (in microns) for each vessel.
}
\description{
Compute the diameter of vessels based on the median radius from
the centroid to vessel polygon edge.
}
