% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ves_fraction_pt_sample}
\alias{ves_fraction_pt_sample}
\title{Fractional area of vessels}
\usage{
ves_fraction_pt_sample(bsf, csf, rsf, vsf, npts = 1e+05, binsize = 200)
}
\arguments{
\item{bsf}{\code{sf} object of bounding box}

\item{csf}{\code{sf} object of cambium line}

\item{rsf}{\code{sf} object of ray line(s)}

\item{vsf}{\code{sf} object of vessel polygons}

\item{npts}{Number of random points to use for sampling}

\item{binsize}{Size of distance bins to away from cambium}
}
\value{
A data.frame with the number of random sample points included in each bin, the number of vessels in each bin, the vessel fraction in each bin, the mean vessel area in each bin, and the vessel density in each bin.
}
\description{
Get fractional area of vessels in distance bins away from cambium based on point sampling.
}
