% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ves_fraction_pt_sample_branch}
\alias{ves_fraction_pt_sample_branch}
\title{Fractional area of vessels from branch cross-sections}
\usage{
ves_fraction_pt_sample_branch(csf, psf, vsf, npts = 1e+05, SaveRandomPts = T)
}
\arguments{
\item{csf}{\code{sf} object of bounding polygon (outer cambium)}

\item{psf}{\code{sf} object of pith polygon}

\item{vsf}{\code{sf} object of vessel polygons}

\item{npts}{Number of random points to use for sampling}

\item{SaveRandomPts}{Logical as to whether the random points should be provided as output.  In this case, the output is a list (see \code{Value}).}
}
\value{
If \code{SaveRandomPts==F}, a data.frame with the vessel fraction and density. If \code{SaveRandomPts==T}, a list where the first element is a data.frame as described above and the second element is an sf-object of the random points used to compute vessel fraction.
}
\description{
Get fractional area of vessels in an entire branch cross section based on point sampling.
}
