% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ves_group_indices}
\alias{ves_group_indices}
\title{Vessel Grouping Index}
\usage{
ves_group_indices(vsf, thresh = 33.26)
}
\arguments{
\item{vsf}{\code{sf} object of vessel polygons}

\item{thresh}{Distance threshold (in microns) to identify grouped vessels. Ideally should be empirically determined as the vessel double wall thickness. Default value is based on ~600 measurements for a set of 10 tree species from the Luquillo Experimental Forest in Puerto Rico.}
}
\value{
A data.frame with grouping indices for vessels included in the sample.
}
\description{
Function for getting vessel grouping index (average number of vessels per group)
and proportion of vessels in groups given a \code{vsf} object. Can be applied to
subsets of the \code{vsf} to, for example, examine changes in grouping index in
different parts of the sample.
}
