% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ves_max_diam}
\alias{ves_max_diam}
\title{Maximum vessel diameter}
\usage{
ves_max_diam(vsf, parallel = TRUE, cores = detectCores(logical = FALSE))
}
\arguments{
\item{vsf}{\code{sf} object of vessel polygons}

\item{parallel}{Logical if processing should be done in parallel
(implemented but of questionably utility).}

\item{cores}{Number of cores to use for parallel processing.}
}
\value{
A vector of maximum diameter (in microns) for each vessel.
}
\description{
Compute the maximum diameter across a vessel based on
2x the longest radius from edge of polygon to centroid.
}
